introduce yarn, rearrange file paths, temporary disable api error module
