+++ bin/www
@@ -2,7 +2,8 @@
 
 const app = require('../index');
 const http = require('http');
-const port = parseInt(process.env.PORT || 3000)
+
+const port = parseInt(process.env.PORT, 10) || 3000;
 app.set('port', port);
 
 const server = http.createServer(app);
