+++ config/config.js
@@ -6,4 +6,4 @@ const config = {
     port: process.env.PORT
 };
 
-export default config;
+module.exports = config;
