+++ server/routes/index.route.js
@@ -10,7 +10,7 @@ const observationTypeRoutes = require('./observation_type.route');
 const observationEvidenceRoutes = require('./observation_evidence.route');
 const observationClusterRoutes = require('./observation_cluster.route');
 const scoreRoutes = require('./score.route');
-const indicatorScoreRoutes = require('./indicator_score.route);
+const indicatorScoreRoutes = require('./indicator_score.route');
 const clusterRoutes = require('./cluster.route');
 const YAML = require('yamljs');
 const swaggerUi = require('swagger-ui-express');
